# go/controlplane/server/openapi.yaml
openapi: "3.0.3"
info:
  title: Purser Control Plane API
  version: "0.2"
  description: |
    Management API for the Purser distributed LLM inference orchestrator.

    Base path: /api/v1

    Authentication: Optional OIDC Bearer token. When PURSER_OIDC_ISSUER is set,
    every request must carry a valid Bearer token in the Authorization header.
    Without that environment variable the server accepts unauthenticated requests.

    Enterprise endpoints (under /api/v1/enterprise/) require a valid
    PURSER_LICENSE_KEY whose offline-verified payload includes the requested
    feature. Without a valid entitlement they return 402 Payment Required.
    The license is verified entirely offline — no phone-home.
servers:
  - url: /api/v1
    description: Purser control-plane API (relative to the server root)
security:
  - BearerAuth: []
  - {}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        OIDC-derived JWT. Required only when PURSER_OIDC_ISSUER is configured.
        Omit on unauthenticated deployments.
  schemas:
...
    Model:
      type: object
      description: A model catalog entry.
      properties:
        id:
          type: string
          example: "llama-3-8b"
        family:
          type: string
          example: "llama"
        architecture:
          type: string
          example: "llama"
        params_total_b:
          type: number
          format: double
          description: Total parameter count in billions.
          example: 8.0
        engine:
          type: string
          example: "vllm"
        spec:
          type: object
          description: Full purserv1.ModelSpec encoded as protojson. Fields are camelCase.
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - family
    Fit:
      type: object
      description: >
        Deployability verdict for a model against the current fleet.
        Present only when the server has a Planner configured.
      properties:
        model_id:
          type: string
          example: "llama-3-8b"
        deployable:
          type: boolean
          example: true
        node_count:
          type: integer
          description: Number of nodes needed for the plan.
          example: 1
        quantization:
          type: string
          description: Chosen quantization level.
          example: "q4_k_m"
        estimated:
          $ref: "#/components/schemas/EstimateRange"
        deficit_gb:
          type: number
          format: double
          description: GB of VRAM/RAM missing when not deployable.
          example: 24.5
        reason:
          type: string
          description: Human-readable explanation when not deployable.
          example: "insufficient VRAM across all ready nodes"
        suggestions:
          type: array
          items:
            type: string
          description: Actionable suggestions when not deployable.
          example: ["add a node with at least 48 GB VRAM"]
    EstimateRange:
      type: object
      description: Estimated throughput range for a deployable model.
      properties:
        decode_min_tok_s:
          type: number
          format: double
          example: 30.0
        decode_max_tok_s:
          type: number
          format: double
          example: 80.0
        prefill_min_tok_s:
          type: number
          format: double
          example: 200.0
        prefill_max_tok_s:
          type: number
          format: double
          example: 600.0
        headroom_gb:
          type: number
          format: double
          description: Remaining VRAM/RAM headroom after the model is loaded.
          example: 12.5
    ModelWithFit:
      allOf:
        - $ref: "#/components/schemas/Model"
        - type: object
          properties:
            fit:
              $ref: "#/components/schemas/Fit"
    CreateModelRequest:
      type: object
      description: >
        A protojson-encoded purserv1.ModelSpec. All field names are lowerCamelCase
        as per the protojson convention. The model_id field (serialised as
        "modelId") is required.
      properties:
        modelId:
          type: string
          example: "llama-3-8b"
        family:
          type: string
          example: "llama"
        architecture:
          type: string
          example: "llama"
        paramsTotalB:
          type: number
          format: double
          example: 8.0
        engine:
          type: string
          example: "vllm"
      required:
        - modelId
      additionalProperties: true

    # --------------------------------------------------------------- plans
...
    Error:
      type: object
      description: >
        Standard error envelope. All non-success responses (except the enterprise
        license gate) use this shape.
      properties:
        error:
          type: string
          description: Machine-readable error kind (e.g. "not_found", "bad_request").
          example: "not_found"
        message:
          type: string
          description: Human-readable description.
          example: "node not found"
      required:
        - error
        - message
...
    ConflictModelInUse:
      type: object
      description: 409 Conflict when a model is referenced by active deployments.
      properties:
        error:
          type: string
          example: "model_in_use"
        message:
          type: string
          example: "model is referenced by one or more active deployments; tear them down first"
        deployments:
          type: array
          items:
            type: string
          description: IDs of the blocking deployments.
          example: ["dep-abc123"]
      required:
        - error
        - message
        - deployments
...
paths:
...
  /models:
    get:
      operationId: listModels
      summary: List the model catalog
      description: >
        Returns all models in the catalog. When a Planner is configured, each
        model is annotated with a "fit" verdict (deployable/infeasible) computed
        against the current fleet, so the UI can render a badge without a second
        round-trip. When no Planner is configured, "fit" is absent.
      security:
        - BearerAuth: []
        - {}
      responses:
        "200":
          description: Model list (with optional fit annotation).
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: "#/components/schemas/ModelWithFit"
                required:
                  - models
        "500":
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      operationId: createModel
      summary: Register a model in the catalog
      description: >
        Accepts a protojson-encoded purserv1.ModelSpec. The model_id field
        (serialised as "modelId") is required. Promoted columns (family,
        architecture, params_total_b, engine) are extracted from the spec.
        The spec is re-marshalled via protojson before storage so it round-trips
        losslessly. Duplicate model_id returns 409.
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelRequest"
            example:
              modelId: "llama-3-8b"
              family: "llama"
              architecture: "llama"
              paramsTotalB: 8
              engine: "vllm"
      responses:
        "201":
          description: Model created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_id:
                    type: string
                    example: "llama-3-8b"
                required:
                  - model_id
        "400":
          description: Empty body, invalid protojson, or missing model_id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A model with this ID already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /models/{id}:
    delete:
      operationId: deleteModel
      summary: Remove a model from the catalog
      description: >
        Guarded delete — never a cascade:
        1. An unknown ID returns 404.
        2. If any non-terminal deployment still references the model, the request
           is refused with 409 model_in_use listing the blocking deployment IDs.
           Tear those deployments down explicitly first.
        3. On success, 204 No Content is returned.
      security:
        - BearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Model ID.
      responses:
        "204":
          description: Model deleted successfully. No content.
        "404":
          description: Model not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Model is pinned by active deployments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictModelInUse"
        "500":
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...